# thermalforge-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermalForge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project_id}/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_v1_projects__project_id__tasks_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel Task",
        "operationId": "cancel_task_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/start": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Start Task",
        "operationId": "start_task_v1_tasks__task_id__start_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "operationId": "retry_task_v1_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/artifacts": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Task Artifacts",
        "operationId": "list_task_artifacts_v1_tasks__task_id__artifacts_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactRead"
                  },
                  "title": "Response List Task Artifacts V1 Tasks  Task Id  Artifacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Stream Task Events",
        "operationId": "stream_task_events_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "follow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Follow"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tasks/{task_id}/clarification": {
      "get": {
        "tags": [
          "clarifications"
        ],
        "summary": "Get Clarification",
        "operationId": "get_clarification_v1_tasks__task_id__clarification_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clarifications"
        ],
        "summary": "Answer Clarification",
        "operationId": "answer_clarification_v1_tasks__task_id__clarification_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationAnswer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tasks/{task_id}/image-manifest": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Task Image Manifest",
        "operationId": "get_task_image_manifest_v1_tasks__task_id__image_manifest_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskImageManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/images/{artifact_id}/content": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Download Task Image",
        "operationId": "download_task_image_v1_tasks__task_id__images__artifact_id__content_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable generated concept image bytes.",
            "content": {
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactKind": {
        "type": "string",
        "enum": [
          "source_document",
          "parsed_document",
          "engineering_brief",
          "thermal_analysis",
          "thermal_design",
          "concept_image",
          "multiview_image",
          "raw_model",
          "normalized_model",
          "report"
        ],
        "title": "ArtifactKind"
      },
      "ArtifactRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "storage_uri": {
            "type": "string",
            "title": "Storage Uri"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "quality_status": {
            "$ref": "#/components/schemas/QualityStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "kind",
          "version",
          "mime_type",
          "sha256",
          "size_bytes",
          "storage_uri",
          "quality_status",
          "created_at"
        ],
        "title": "ArtifactRead"
      },
...
      "ClarificationAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "ClarificationAnswer"
      },
      "ClarificationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "field_key": {
            "type": "string",
            "title": "Field Key"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "answered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "field_key",
          "question",
          "answer",
          "created_at",
          "answered_at"
        ],
        "title": "ClarificationRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QualityStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "QualityStatus"
      },
...
      "TaskCreate": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 10000,
            "title": "Prompt",
            "default": ""
          }
        },
        "type": "object",
        "title": "TaskCreate"
      },
      "TaskImageAsset": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "concept_image",
              "multiview_image"
            ],
            "title": "Kind"
          },
          "view_id": {
            "type": "string",
            "title": "View Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "mime_type": {
            "type": "string",
            "const": "image/png",
            "title": "Mime Type"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "provider_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Model"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "kind",
          "view_id",
          "url",
          "mime_type",
          "sha256",
          "size_bytes"
        ],
        "title": "TaskImageAsset"
      },
      "TaskImageManifest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/TaskImageAsset"
            },
            "type": "array",
            "title": "Images"
          },
          "notice": {
            "type": "string",
            "title": "Notice",
            "default": "概念图用于方案沟通，不是 CAD、CFD、FEA 或制造验证结果。"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "TaskImageManifest"
      },
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "stage",
          "prompt",
          "idempotency_key",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "created",
          "uploaded",
          "parsing",
          "awaiting_input",
          "briefing",
          "thermal_analysis",
          "concept_imaging",
          "multiview_imaging",
          "multiview_review",
          "modeling",
          "model_review",
          "ready",
          "failed",
          "cancelled"
        ],
        "title": "TaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}